feat(validation): cache solc input
Within the worker (`workerState`) we retain the last input to solc
calculated by the preprocessing steps. If the same uri with the same
solc version and imports comes in again, we reuse the same input with an
updated content for the one changed file.

This gives a performance improvement (the preprocessing is the larger
part of validation in big codebases) for the most common case which is
lots of changes in a single file.